[ADD] SavePlayerScore function
on player death, writes their score to a score.txt file

+++ Server/Interface/Include/level.hpp
@@ -196,12 +196,6 @@ template <typename T> class Level : virtual public ILevel<T> {
                                 }
                                 playerHealth1.value()->health -= 1;
                             }
-                            if (playerHealth1.value()->health <= 0) {
-                                if (std::find(entitiesToRemove.begin(), entitiesToRemove.end(),
-                                        entityId1) == entitiesToRemove.end()) {
-                                    entitiesToRemove.push_back(entityId1);
-                                }
-                            }
                             r_type::net::Message<TypeMessage> updLivesMsg;
                             updLivesMsg.header.id = TypeMessage::UpdateInfoBar;
                             updLivesMsg << server->UpdateInfoBar(entityId1);
@@ -310,12 +304,7 @@ template <typename T> class Level : virtual public ILevel<T> {
                                 }
                                 playerHealth2.value()->health -= 1;
                             }
-                            if (playerHealth2.value()->health <= 0) {
-                                if (std::find(entitiesToRemove.begin(), entitiesToRemove.end(),
-                                        entityId2) == entitiesToRemove.end()) {
-                                    entitiesToRemove.push_back(entityId2);
-                                }
-                            }
+
                             r_type::net::Message<TypeMessage> updLivesMsg;
                             updLivesMsg.header.id = TypeMessage::UpdateInfoBar;
                             updLivesMsg << server->UpdateInfoBar(entityId2);
@@ -324,7 +313,7 @@ template <typename T> class Level : virtual public ILevel<T> {
                                 updLivesMsg);
                             if (playerHealth2.value()->health <= 0) {
                                 if (std::find(entitiesToRemove.begin(), entitiesToRemove.end(),
-                                        entityId2) == entitiesToRemove.end()) {
+                                        entityId1) == entitiesToRemove.end()) {
                                     entitiesToRemove.push_back(entityId2);
                                 }
                             }
@@ -434,6 +423,9 @@ template <typename T> class Level : virtual public ILevel<T> {
             msg.header.id = TypeMessage::DestroyEntityMessage;
             msg << entityId;
             server->MessageAllClients(msg);
+            if (auto playerComponent = componentManager.getComponent<PlayerComponent>(entityId)) {
+                continue;
+            }
             componentManager.removeEntityFromAllComponents(entityId);
             entityManager.removeEntity(entityId);
         }
